requirements.txt
kafka-python
pymongo
flask
streamlit
requests